This is a branch dedicated to distributed flask-ifot middleware. The frontend is limited to API calls only. See other git hub (to be added here) for the frontend.
Setup is as follows:
This is for the cluster of nodes (manager/worker).
Frontend (other git) is for service broker.

Issues encountered are CORS and CORB, where ajax calls from a different origin are blocked (for security reasons).
Workaround, just allow all requests from all sources since this is for development.

Other issue:
Even in distributed, the collection and distribution of chunks is serial. So it takes the same time as the single node baseline (relatively).
TODO: Will need to paralellize the distribution.